move dev tool to subdirectory
