+++ apps/dev-tool/next-env.d.ts
