+++ apps/dev-tool/public/favicon.ico
